feat(neuron-kit): SCENARIO_WIRE_PARITY_001 — shared camelCase wire + f64; cross-leg byte gate live
Per Bob's serde-precedent ruling and the three-leg analysis: (1)
ScenarioProfile's Rust wire keys are now the Swift Codable camelCase
vocabulary (serde rename_all + explicit profileID — Swift's property is
capital-ID; rename_all alone would emit profileId), with a wire-key
test; (2) scoringBreakdown/preferenceWeights switch Float/f32 →
Double/f64 in BOTH legs — JSON numbers are f64-natured and
shortest-roundtrip f64 text is the unique point where Swift, Rust, and
the coming Go leg's standard encoders all emit identical bytes (f32 is
a three-way disagreement: NSNumber 9-sig-digits vs serde f64-widening
vs Go shortest-f32); (3) the conformance artifact's scenario_profile
section re-recorded and the Rust verifier upgraded from per-leg
round-trip to a BYTE comparison against canonicalJson — green on first
verify; (4) the stale 'bit-identical… kebab-case to snake_case' header
replaced with the now-true wire-parity statement; INTERFACE § 1/§ 2
types updated; computational cost of f64 here: nil (scalar f64 ==
scalar f32 on modern cores; values never in SIMD paths; wire text
shrinks).

Test Verification Log: NeuronKit swift test 172 / 29 exit 0; neuron-kit
cargo test 162 + 1 integration exit 0 (incl. the new cross-leg byte
gate + wire-key test); clippy -D warnings exit 0; fmt clean.

+++ packages/kits/NeuronKit/Sources/NeuronKit/ScenarioProfile.swift
@@ -50,12 +50,12 @@ public struct ScenarioProfile: Sendable, Equatable, Codable {
     /// Per-signal breakdown of the tournament's scoring. Keys are the
     /// signal names (e.g., "averageReward"); values are the numeric
     /// score the tournament observed.
-    public let scoringBreakdown: [String: Float]
+    public let scoringBreakdown: [String: Double]
 
     /// User-tunable preference weights applied on top of the scoring
     /// breakdown. Spec § 4.6 — saved here so future runs can replay
     /// the preference signal without rebuilding it.
-    public let preferenceWeights: [String: Float]
+    public let preferenceWeights: [String: Double]
 
     /// Wall-clock at which the profile was saved. Carried by the
     /// caller; the substrate stores it as ISO8601 TEXT per the fleet
@@ -72,8 +72,8 @@ public struct ScenarioProfile: Sendable, Equatable, Codable {
         profileID: UUID = UUID(),
         name: String,
         framingParameters: [String: String],
-        scoringBreakdown: [String: Float],
-        preferenceWeights: [String: Float],
+        scoringBreakdown: [String: Double],
+        preferenceWeights: [String: Double],
         createdAt: Date,
         trainingEligible: Bool = false
     ) {
